WI-5 Phase 5.1: Extract platform-neutral completion logic into shared/
Extract platform-neutral modules from completions/ and documentdbQuery/
into src/webviews/documentdbQuery/shared/:

Moved (git mv, preserves history):
- completionKnowledge.ts — KEY_POSITION_OPERATORS, placeholder constants
- snippetUtils.ts — stripOuterBraces(), escapeSnippetDollars()
- cursorContext.ts — detectCursorContext(), CursorContext type
- extractQuotedKey.ts — extractQuotedKey() for hover range

Extracted (new files, platform-neutral parts from mixed modules):
- sortPrefixes.ts — getOperatorSortPrefix(), getCategoryLabel()
  (from mapCompletionItems.ts, removes duplication with scratchpad)
- typeSuggestionData.ts — TypeSuggestionDef, TYPE_SUGGESTIONS map,
  getTypeSuggestionDefs() (from typeSuggestions.ts)
- jsGlobalDefs.ts — JsGlobalDef, JS_GLOBALS array (from jsGlobals.ts)

Consolidated:
- ScratchpadCompletionItemProvider.getVscodeOperatorSortPrefix() now
  wraps shared getOperatorSortPrefix() instead of duplicating the logic
- All imports updated across both scratchpad and webview consumers
- Barrel index.ts for shared/ and updated completions/index.ts

No behavioral changes — purely structural refactoring.
All 1118 tests pass.
