+++ src/webviews/documentdbQuery/completions/createCompletionItems.ts
@@ -19,9 +19,9 @@ import {
 // eslint-disable-next-line import/no-internal-modules
 import type * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api';
 import { getCompletionContext } from '../completionStore';
-import { type CursorContext } from '../cursorContext';
+import { type CursorContext } from '../shared/cursorContext';
 import { EditorType } from '../languageConfig';
-import { KEY_POSITION_OPERATORS } from './completionKnowledge';
+import { KEY_POSITION_OPERATORS } from '../shared/completionKnowledge';
 import { createJsGlobalCompletionItems } from './jsGlobals';
 import { mapFieldToCompletionItem, mapOperatorToCompletionItem } from './mapCompletionItems';
 import { createTypeSuggestions } from './typeSuggestions';
@@ -62,7 +62,7 @@ export interface CreateCompletionItemsParams {
 
 // KEY_POSITION_OPERATORS is imported from ./completionKnowledge
 // Re-export for backwards compatibility and testing
-export { KEY_POSITION_OPERATORS } from './completionKnowledge';
+export { KEY_POSITION_OPERATORS } from '../shared/completionKnowledge';
 
 /**
  * Returns the completion meta tags appropriate for the given editor type.
