+++ src/webviews/documentdbQuery/registerLanguage.ts
@@ -27,10 +27,10 @@
 // eslint-disable-next-line import/no-internal-modules
 import type * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api';
 import { getCompletionContext } from './completionStore';
-import { detectCursorContext } from './cursorContext';
+import { detectCursorContext } from './shared/cursorContext';
 import { createCompletionItems } from './documentdbQueryCompletionProvider';
 import { getHoverContent } from './documentdbQueryHoverProvider';
-import { extractQuotedKey } from './extractQuotedKey';
+import { extractQuotedKey } from './shared/extractQuotedKey';
 import { isCursorInsideString } from './isCursorInsideString';
 import { LANGUAGE_ID, parseEditorUri } from './languageConfig';
