+++ src/webviews/documentdbQuery/shared/completionKnowledge.ts
