+++ src/webviews/documentdbQuery/shared/extractQuotedKey.ts
