+++ src/webviews/documentdbQuery/tdd/completionBehavior.test.ts
@@ -27,7 +27,7 @@
 // eslint-disable-next-line import/no-internal-modules
 import type * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api';
 import { clearAllCompletionContexts, setCompletionContext } from '../completionStore';
-import { type CursorContext } from '../cursorContext';
+import { type CursorContext } from '../shared/cursorContext';
 import { createCompletionItems } from '../documentdbQueryCompletionProvider';
 import { EditorType } from '../languageConfig';
